fixed authentication

+++ public/index.php
@@ -19,14 +19,14 @@ $router->post('/logout', 'AuthController@logout')->middleware('auth');
 
 $router->get('/tasks', 'TaskController@index')->name('tasks.index');
 
-$router->group(['prefix' => 'tasks', 'middleware' => ['auth']], function (RouteProxyInterface $router) {
+$router->group(['prefix' => 'tasks'], function (RouteProxyInterface $router) {
     $router->get('/create', 'TaskController@create')->name('tasks.create');
     $router->post('/', 'TaskController@store')->name('tasks.store');
-    $router->delete('{id}', 'TaskController@delete')->name('tasks.destroy');
-    $router->get('{id}/edit', 'TaskController@edit')->name('tasks.edit');
-    $router->put('{id}', 'TaskController@update')->name('tasks.update');
-    $router->post('{id}/finish', 'TaskController@finish')->name('tasks.finsih');
-    $router->post('{id}/unfinish', 'TaskController@unfinish')->name('tasks.unfinish');
+    $router->delete('{id}', 'TaskController@delete')->name('tasks.destroy')->middleware(['admin']);
+    $router->get('{id}/edit', 'TaskController@edit')->name('tasks.edit')->middleware(['admin']);
+    $router->put('{id}', 'TaskController@update')->name('tasks.update')->middleware(['admin']);
+    $router->post('{id}/finish', 'TaskController@finish')->name('tasks.finsih')->middleware(['admin']);
+    $router->post('{id}/unfinish', 'TaskController@unfinish')->name('tasks.unfinish')->middleware(['admin']);
 });
 
 $app->handle($request);
